refactor: convert twomap.rs file into a cmap2 module
