+++ honeycomb-core/src/lib.rs
@@ -29,8 +29,8 @@
 // ------ MODULE DECLARATIONS
 mod attributes;
 mod cells;
+mod cmap2;
 mod spatial_repr;
-mod twomap;
 
 #[cfg(feature = "utils")]
 pub mod utils;
@@ -50,10 +50,10 @@ pub use cells::{
     orbits::{Orbit2, OrbitPolicy},
     DartIdentifier, NULL_DART_ID,
 };
+pub use cmap2::{CMap2, CMapError};
 pub use spatial_repr::{
     coords::{Coords2, CoordsError},
     vector::Vector2,
     vertex::Vertex2,
     CoordsFloat, FloatType,
 };
-pub use twomap::{CMap2, CMapError};
